requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
stripe==7.2.0
qrcode==7.4.2
Pillow==10.1.0
reportlab==4.0.7
python-dotenv==1.0.0
Flask-Mail==0.9.1
